# Ensure ShapedArray.shape is always a tuple of builtins integers