--- a/jax/interpreters/masking.py
+++ b/jax/interpreters/masking.py
@@ -88,6 +88,10 @@ def mask_subtrace(master, in_vals, shape_exprs):
   out_vals, out_shapes = unzip2((t.val, t.shape_expr) for t in out_tracers)
   yield out_vals, out_shapes
 
+def to_index(x):
+  """Like operator.index, but allowing polymorphic dimensions."""
+  # not implemented as `Poly.__index__`, since operator.index only allows ints
+  return x if type(x) is Poly else op.index(x)
 
 def ensure_poly(p):
   if isinstance(p, Poly):