@shoyer 2020-01-29
I tried updating all the various shape rules in lax.py to always return builtin integers instead of NumPy scalars, but there are quite a few of them and it got pretty annoying.

I do not measure any noticeable change in performance in my microbenchmark:
Before:
```
In [6]: %timeit x + 1
179 µs ± 5.06 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```
After:
```
In [9]: %timeit x + 1
180 µs ± 3.99 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

I guess the `shape` gets cached?