@shoyer 2020-01-29
Also tested on `x = lax.add(1, 1)`:
Before:
```
In [9]: %timeit lax.add(x, x)
51.5 µs ± 486 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```
After:
```
In [8]: %timeit lax.add(x, x)
51.8 µs ± 403 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```